+++ lib/main.dart
@@ -1,6 +1,6 @@
 /// TodoPod — privacy-first todo list with Solid Pod storage main entry point.
 ///
-// Time-stamp: <Wednesday 2026-04-29 15:16:11 +1000 Graham Williams>
+// Time-stamp: <Thursday 2026-04-30 12:14:09 +1000 Graham Williams>
 ///
 /// Copyright (C) 2026, Togaware Pty Ltd
 ///
@@ -34,18 +34,54 @@ import 'package:flutter/material.dart';
 
 import 'package:provider/provider.dart';
 import 'package:solidui/solidui.dart';
+import 'package:window_manager/window_manager.dart';
 
 import 'package:todopod/app_scaffold.dart';
 import 'package:todopod/constants/app.dart';
 import 'package:todopod/services/app_provider.dart';
 
-void main() {
+// 20260402 gjw Below is the main entry point for the application.  For main()
+// we require [async] because we asynchronously [await] the window manager as
+// below. Often, `main()` will include just the call [runApp].
+
+void main() async {
+  // 20260402 gjw Optionally for development we utilise [debugPrint] to trace
+  // execution, and note that the output is not shown on a `--release`. To
+  // quieten the `--debug` running we can globally remove [debugPrint] messages
+  // by mapping it to null (no op).
+  //
+  // debugPrint = (String? message, {int? wrapWidth}) {
+  //   null;
+  // };
+
+  // 20260402 gjw We want to ensure Flutter bindings are initialized for async
+  // operations particularly to set the Linux desktop window [title] as we do
+  // below.
+
   WidgetsFlutterBinding.ensureInitialized();
-
-  // Initialise SolidUI security key manager so it can automatically prompt
-  // the user for their security key whenever solidpod needs it.
-
   SolidSecurityKeyCentralManager.instance;
+  if (isDesktop) {
+    await windowManager.ensureInitialized();
+
+    // 20260402 gjw For our desktop app we tune various window oriented
+    // settings.
+
+    const windowOptions = WindowOptions(
+      title: appTitle,
+      minimumSize: Size(500, 800),
+      backgroundColor: Colors.transparent,
+      skipTaskbar: false,
+      titleBarStyle: TitleBarStyle.normal,
+    );
+
+    // 20260402 gjw Now we await the window being shown and receiving the focus,
+    // to then proceed to run the app.
+
+    await windowManager.waitUntilReadyToShow(windowOptions, () async {
+      await windowManager.show();
+      await windowManager.focus();
+    });
+  }
 
   runApp(
     ChangeNotifierProvider(
@@ -94,7 +130,7 @@ class _TodoPodAppState extends State<TodoPodApp> {
       home: SolidLogin(
         required: false,
         appDirectory: appDirectory,
-        title: appName.toUpperCase(),
+        title: appTitle.replaceAll(' - ', '\n'),
         image: const AssetImage('assets/images/app_image.jpg'),
         logo: const AssetImage('assets/images/app_icon.png'),
         link: 'https://github.com/gjwgit/todopod',